+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+#### Fixed
+* **\[processor\]** Ensure that a react component with a parameterized type will add the parameterized type
+  to the builder step interfaces to avoid compilation errors if the type parameter is used by the prop methods.
+
 ### [v0.58](https://github.com/react4j/react4j/tree/v0.58) (2018-04-03)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.57...v0.58)